// src/shared/enums.ts
export enum eModalIds {
  PeopleFilters = 'people_filters',
  ModalSaved = 'modal_saved',
  Shva = 'shva_year_s',
  EducationOrg = 'education_org',
  EducationSpec = 'education_spec',
  WorkSpec = 'work_spec',
  ProfCan = 'prof_can',
  ProfWant = 'prof_want',
  ProfTeach = 'prof_teach',
  ProfLearn = 'prof_learn',
  Interests = 'interests',
  AchievementsHistory = 'achievements.history',
}

export const eFilterStats = [
  'shva_year_s',
  'education_org_active',
  'education_spec_active',
  'work_spec_active',
  'prof_active',
  'interests_active',
  'achievements.history',
]

export enum eViewIds {
  Block = 'block',
  NotLoaded = 'notloaded',
  Loader = 'loader',
  Main = 'main',
}

export enum ePanelIds {
  Home = 'home',
  People = 'people',
  Match = 'match',
  Map = 'map',
  Stats = 'stats',
  Chats = 'chats',
  Block = 'block',
  NotLoaded = 'notloaded',
  Loader = 'loader',
}